pyproject.toml
[tool.poetry]
name = "Decide"
version = "0.1.0"
description = "Decide what to build next"
authors = ["Mathieu <contact@mathieubellon.fr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.1"
Django = "^5.0.3"
django-urls = "^1.1.3"
djangorestframework = "^3.15.1"
uvicorn = "^0.29.0"
dj-database-url = "^2.1.0"
django-filter = "^24.2"
markdown = "^3.6"
whitenoise = "^6.6.0"
psycopg2 = "^2.9.9"
gunicorn = "^21.2.0"
faker = "^24.4.0"
django-extensions = "^3.2.3"
django-allauth = "^0.61.1"
django-environ = "^0.11.2"
pydantic = "^2.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
